+++ src/Native/index.ts
@@ -1,0 +1,2 @@
+export { NativeInput } from "./NativeInput";
+export { NativeInputRange } from "./NativeInputRange";
